# specs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scopegate API contract",
    "version": "1.0.0",
    "description": "Executable local product contract. Authority and trust boundaries are defined in docs/SECURITY.md."
  },
  "servers": [
    {
      "url": "http://localhost:5187",
      "description": "Local same-origin workspace"
    },
    {
      "url": "http://localhost:8457",
      "description": "Local API transport"
    }
  ],
  "paths": {
...
    "/api/v1/organizations/{organization_id}/memberships/{membership_id}/projects/{project_id}/grants": {
      "get": {
        "operationId": "getGrants",
        "summary": "getGrants",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantPage"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Membership access aggregate version",
                "schema": {
                  "type": "string",
                  "pattern": "^\"v[1-9][0-9]*\"$"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/MembershipId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ]
      },
      "patch": {
        "operationId": "applyGrantDiff",
        "summary": "applyGrantDiff",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantResult"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Membership access aggregate version",
                "schema": {
                  "type": "string",
                  "pattern": "^\"v[1-9][0-9]*\"$"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "412": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/MembershipId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantDiff"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organizations/{organization_id}/access-decisions": {
      "post": {
        "operationId": "evaluateOwnAccess",
        "summary": "Evaluate current actor only; result never substitutes for check at use",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDecision"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessDecisionRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/projects/{project_id}/report-configs": {
      "post": {
        "operationId": "createReportConfig",
        "summary": "createReportConfig",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportConfig"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/projects/{project_id}/report-configs/{report_config_id}": {
      "get": {
        "operationId": "getReportConfig",
        "summary": "getReportConfig",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportConfig"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/ReportId"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "scopegate_session"
      },
      "publisherBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated publisher machine credential; catalog write only"
      },
      "platformBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated platform operations credential; administrative provisioning only, no implicit resource use"
      }
    },
    "parameters": {
      "OrganizationId": {
        "name": "organization_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "MembershipId": {
        "name": "membership_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "ReportId": {
        "name": "report_config_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string",
          "maxLength": 512
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 25
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^\"v[1-9][0-9]*\"$"
        }
      },
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 32
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "correlation_id": {
                "type": "string",
                "format": "uuid"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "required": [
              "code",
              "message",
              "correlation_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Grant": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "resource_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "revoked"
            ]
          }
        },
        "required": [
          "resource_id",
          "state"
        ]
      },
      "GrantDiff": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "add": {
            "type": "array",
            "maxItems": 100,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "remove": {
            "type": "array",
            "maxItems": 100,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500
          }
        },
        "required": [
          "add",
          "remove",
          "reason"
        ]
      },
      "GrantResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "membership_id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "access_version": {
            "type": "integer",
            "minimum": 1
          },
          "changed_grants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Grant"
            },
            "maxItems": 200
          }
        },
        "required": [
          "membership_id",
          "project_id",
          "access_version",
          "changed_grants"
        ]
      },
...
      "ReportCreate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "resource_ids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "name",
          "resource_ids"
        ]
      },
      "ReportConfig": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "resource_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "id",
          "organization_id",
          "project_id",
          "name",
          "resource_ids"
        ]
      },
      "AccessDecisionRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "resource_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "project_id",
          "resource_id"
        ]
      },
      "AccessDecision": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "enum": [
              "allowed",
              "organization_inactive",
              "membership_inactive",
              "membership_expired",
              "project_inactive",
              "entitlement_inactive",
              "resource_unpublished",
              "grant_missing"
            ]
          }
        },
        "required": [
          "allowed",
          "reason"
        ]
      },
...
      "GrantPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "membership_id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "access_version": {
            "type": "integer",
            "minimum": 1
          },
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/Grant"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "membership_id",
          "project_id",
          "access_version",
          "items",
          "next_cursor"
        ],
        "description": "Cursor binds membership access_version. If version changes while paging, return 412 rather than mix grant snapshots."
      },
...
    },
    "responses": {
      "Error": {
        "description": "Stable machine-readable error. Inaccessible objects use 404; known insufficient management scope uses 403.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}